% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{AddMapsInput}
\alias{AddMapsInput}
\title{AddMapsInput}
\usage{
AddMapsInput(hierarchies, data = NULL)
}
\arguments{
\item{hierarchies}{hierarchies}

\item{data}{data}
}
\description{
Brukes til å generere feil
}
\keyword{internal}
