% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{AddNonExistingCode}
\alias{AddNonExistingCode}
\title{AddNonExistingCode}
\usage{
AddNonExistingCode(hierarchies, rowSelect = NULL, inputInOutput = TRUE)
}
\arguments{
\item{hierarchies}{hierarchies}

\item{rowSelect}{rowSelect}

\item{data}{data}
}
\description{
AddNonExistingCode
}
\keyword{internal}
