% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/As_TsparseMatrix.R
\name{As_TsparseMatrix}
\alias{As_TsparseMatrix}
\title{Transform to TsparseMatrix/dgTMatrix}
\usage{
As_TsparseMatrix(from, do_drop0 = TRUE)
}
\arguments{
\item{from}{A matrix}

\item{do_drop0}{whether to run \code{drop0}}
}
\value{
A matrix. Virtual class is \code{TsparseMatrix}. Class \code{dgTMatrix} expected.
}
\description{
To implement adaption needed after Matrix ver. 1.4-2 since
\code{as(from, "dgTMatrix")} no longer allowed.
}
\note{
\code{Matrix:::.as.via.virtual} in development version of package \code{Matrix} (date 2022-08-13) used to generate code.
}
