% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{combine_formulas}
\alias{combine_formulas}
\title{Function for linking tables defined using model formulas}
\usage{
combine_formulas(lof, simplify = TRUE, env = parent.frame())
}
\arguments{
\item{lof}{list or vector of formulas to be linked}

\item{simplify}{logical value, default TRUE. Determines whether the formula
should be simplified before output or not.}

\item{env}{the environment for the output formula}
}
\value{
model formula
}
\description{
Function for linking tables defined using model formulas
}
\examples{
lof1 <- c(~a+b, ~a:c, ~c*d)
combine_formulas(lof1, simplify = TRUE)
combine_formulas(lof1, simplify = FALSE)
}
