% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\name{combine_table_formulas}
\alias{combine_table_formulas}
\title{Function for linking tables defined using model formulas}
\usage{
combine_table_formulas(lof, simplify = TRUE)
}
\arguments{
\item{lof}{list or vector of formulas to be linked}

\item{simplify}{logical value, default TRUE. Determines whether the formula
should be simplified before output or not.}
}
\value{
model formula
}
\description{
Function for linking tables defined using model formulas
}
\examples{
lof1 <- c(~a+b, ~a:c, ~c*d)
combine_table_formulas(lof1, simplify = TRUE)
combine_table_formulas(lof1, simplify = FALSE)
}
