% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToMatrix.R
\name{CrossCodeFrames}
\alias{CrossCodeFrames}
\title{Cross codes in data frames}
\usage{
CrossCodeFrames(codeFrame1, codeFrame2, useMatrixToDataFrame = TRUE)
}
\arguments{
\item{codeFrame1}{data frame}

\item{codeFrame2}{data frame}

\item{useMatrixToDataFrame}{useMatrixToDataFrame}
}
\value{
data frame
}
\description{
Cross codes in data frames
}
\author{
Øyvind Langsrud
}
\keyword{internal}
