% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{CrossDataDummyHierarchies}
\alias{CrossDataDummyHierarchies}
\title{CrossDataDummyHierarchies}
\usage{
CrossDataDummyHierarchies(dataDummyHierarchies, codeFrames = NULL,
  makeDimnames = FALSE, useMatrixToDataFrame = TRUE)
}
\arguments{
\item{dataDummyHierarchies}{dataDummyHierarchies}

\item{codeFrames}{codeFrames}

\item{makeDimnames}{makeDimnames}
}
\description{
CrossDataDummyHierarchies
}
\keyword{internal}
