% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{CrossDataDummyHierarchy}
\alias{CrossDataDummyHierarchy}
\title{CrossDataDummyHierarchy}
\usage{
CrossDataDummyHierarchy(dataDummyHierarchy1, dataDummyHierarchy2 = NULL,
  codeFrame1 = NULL, codeFrame2 = NULL, makeDimnames = FALSE,
  useMatrixToDataFrame = TRUE)
}
\arguments{
\item{dataDummyHierarchy1}{dataDummyHierarchy1}

\item{dataDummyHierarchy2}{dataDummyHierarchy2}

\item{codeFrame1}{codeFrame1}

\item{codeFrame2}{codeFrame2}

\item{makeDimnames}{makeDimnames}
}
\description{
CrossDataDummyHierarchy
}
\keyword{internal}
