% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatternMatching.R
\name{crossMerge}
\alias{crossMerge}
\title{crossMerge}
\usage{
crossMerge(ind1, ind2, x, y, useMatrixToDataFrame = TRUE)
}
\arguments{
\item{ind1}{ind1}

\item{ind2}{ind2}

\item{x}{x}

\item{y}{y}
}
\description{
crossMerge
}
\keyword{internal}
