% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToMatrix.R
\name{DataFrameToMatrix}
\alias{DataFrameToMatrix}
\title{DataFrameToMatrix}
\usage{
DataFrameToMatrix(x)
}
\arguments{
\item{x}{data frame}
}
\value{
matrix
}
\description{
DataFrameToMatrix
}
\author{
Øyvind Langsrud
}
\keyword{internal}
