% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_multiple_fun.R
\name{fix_fun_amf}
\alias{fix_fun_amf}
\title{Fix \code{fun} parameter to \code{aggregate_multiple_fun}}
\usage{
fix_fun_amf(fun)
}
\arguments{
\item{fun}{fun}
}
\value{
fun
}
\description{
Fix \code{fun} parameter to \code{aggregate_multiple_fun}
}
\examples{
identical(fix_fun_amf("median"), c(median = median))

identical(fix_fun_amf(c("sum", "median")), c(sum = sum, median = median))

ff <- c("sum", "median", "cor")
names(ff) <- c("", NA, "Correlation")
identical(fix_fun_amf(ff), c(sum, median = median, Correlation = cor))

identical(fix_fun_amf(structure("median", names = "")), fix_fun_amf(median))
}
\keyword{internal}
