% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_multiple_fun.R
\name{fix_vars_amf}
\alias{fix_vars_amf}
\title{Fix \code{vars} parameter to \code{aggregate_multiple_fun}}
\usage{
fix_vars_amf(vars, multi_sep)
}
\arguments{
\item{vars}{vars}

\item{multi_sep}{multi_sep}
}
\description{
Fix \code{vars} parameter to \code{aggregate_multiple_fun}
}
\examples{
f <- fix_vars_amf

f(c("ant", "y", median = "ant", median = "y", d1 = "ant"))

v1 <- list(sum = "a", sum = "w", q = c("a", "w"), snitt = c("b", "w"))
v2 <- list(c(FUN = "sum", "a"), c(FUN = "sum", "w"), c(FUN = "q", "a", "w"), 
           c(FUN = "snitt", "b", "w"))
v3 <- list(sum = "a", sum = "w", q = c(name = "a:w_q", "a", "w"), 
           `b:w_snitt` = list(snitt = c("b", "w")))
v4 <- list(c(name = "a_sum", FUN = "sum", "a"), 
           c(name = "w_sum", FUN = "sum", "w"), 
           c(name = "a:w_q", FUN = "q", "a", "w"), 
           c(name = "b:w_snitt", FUN = "snitt", "b", "w"))
v5 <- list(a_sum = c(FUN = "sum", "a"), 
           w_sum = c(FUN = "sum", "w"), 
           `a:w_q` = c(FUN = "q", "a", "w"), 
           `b:w_snitt` = c(FUN = "snitt", "b", "w"))

identical(f(v1), f(v2))
identical(f(v1), f(v3))
identical(f(v1), f(v4))
identical(f(v1), f(v5))

identical(f(v1), f(f(v1)))
identical(f(v1), v4)
}
\keyword{internal}
