% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{FixHierarchy}
\alias{FixHierarchy}
\title{FixHierarchy}
\usage{
FixHierarchy(hi, hierarchyVarNames = c(mapsFrom = "from", mapsTo = "to",
  sign = "sign", level = "level"))
}
\arguments{
\item{hi}{hi}

\item{hierarchyVarNames}{hierarchyVarNames}
}
\description{
FixHierarchy
}
\keyword{internal}
