% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToMatrix.R
\name{ForceCharacterDataFrame}
\alias{ForceCharacterDataFrame}
\title{ForceCharacterDataFrame}
\usage{
ForceCharacterDataFrame(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame
}
\description{
ForceCharacterDataFrame
}
\author{
Øyvind Langsrud
}
\keyword{internal}
