% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToMatrix.R
\name{ForceFactorDataFrame}
\alias{ForceFactorDataFrame}
\title{ForceFactorDataFrame}
\usage{
ForceFactorDataFrame(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame
}
\description{
ForceFactorDataFrame
}
\keyword{internal}
