% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_utils.R
\docType{data}
\name{formula_utils}
\alias{formula_utils}
\title{Functions for formula manipulation}
\description{
Functions for formula manipulation
}
\details{
\itemize{
\item \code{\link{combine_formulas}}: Combine formulas
\item \code{\link{formula_from_vars}}:  Generate model formula by specifying which variables have totals or not
\item \code{\link{substitute_formula_vars}}: Replace variables in formula with sum of other variables
}
}
