% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHierFormula.R
\name{HierarchicalGroups2}
\alias{HierarchicalGroups2}
\title{Finding hierarchical variable groups}
\usage{
HierarchicalGroups2(x)
}
\arguments{
\item{x}{Matrix or data frame containing the variables}
}
\value{
List containing the groups.
}
\description{
As HierarchicalGroups() with eachName = TRUE, but output belonging to same mainName are combined.
}
\author{
Øyvind Langsrud
}
\keyword{internal}
