% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalGroups.R
\name{HierarchicalGroups3}
\alias{HierarchicalGroups3}
\title{Finding hierarchical variable groups}
\usage{
HierarchicalGroups3(x)
}
\arguments{
\item{x}{Matrix or data frame containing the variables}
}
\value{
List containing the groups.
}
\description{
As HierarchicalGroups() with eachName = FALSE, but output belonging to same mainName are combined.
}
\author{
Øyvind Langsrud
}
\keyword{internal}
