% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoHierarchies.R
\name{Hierarchy2Formula}
\alias{Hierarchy2Formula}
\alias{Formula2Hierarchy}
\title{Hierarchy2Formula}
\usage{
Hierarchy2Formula(x, hierarchyVarNames = c(mapsFrom = "mapsFrom", mapsTo
  = "mapsTo", sign = "sign", level = "level"))

Formula2Hierarchy(s)
}
\arguments{
\item{x}{Data frame with to-from coded hierarchy}

\item{hierarchyVarNames}{Variable names in the hierarchy tables as in \code{\link{HierarchyFix}}.}

\item{s}{Character vector of formulas written with =, - and +.}
}
\value{
See Arguments
}
\description{
Conversion between to-from coded hierarchy and formulas written with =, - and +.
}
\examples{
x <- SSBtoolsData("sprt_emp_geoHier")
s <- Hierarchy2Formula(x)
Formula2Hierarchy(s)

}
\author{
Øyvind Langsrud
}
