% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_default.R
\name{inc_default}
\alias{inc_default}
\title{Default progress indicator function}
\usage{
inc_default(i = 0L, n = 0L, steps = 25L, dot = ".")
}
\arguments{
\item{i}{i in "i out of n"}

\item{n}{n in "i out of n"}

\item{steps}{Number of dots to print}

\item{dot}{dot}
}
\description{
Default progress indicator function
}
\examples{
for (i in 1:5) inc_default(i, 5)
cat("\n")
for (i in 1:100) inc_default(i, 100)
cat("\n")
for (i in 1:1000) inc_default(i, 1000)
cat("\n")
for (i in 1:1000) inc_default(i, 1000, steps = 10)
cat("\n")
for (i in 1:10) inc_default()
cat("\n")
}
\keyword{internal}
