% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapHierarchiesToData.R
\name{MapHierarchiesToData}
\alias{MapHierarchiesToData}
\title{Add variables to dataset based on hierarchies}
\usage{
MapHierarchiesToData(data, hierarchies, ...)
}
\arguments{
\item{data}{A data frame containing variables with names matching the names of the hierarchies.}

\item{hierarchies}{List of hierarchies on the same form as input to \code{\link{AutoHierarchies}}}

\item{...}{Further parameters sent to \code{\link{Hierarchies2Vars}}}
}
\value{
Input \code{data} with extra Variables
}
\description{
Uses \code{\link{Hierarchies2Vars}} to transform hierarchies, followed by mapping to the dataset.
}
\examples{

# Examples similar those from Hierarchies2Vars

z <- SSBtoolsData("sprt_emp_withEU")
yearFormula <- c("y_14 = 2014", "y_15_16 = y_all - y_14", "y_all = 2014 + 2015 + 2016")
geoDimList <- FindDimLists(z[, c("geo", "eu")], total = "Europe")[[1]]
ageHier <- SSBtoolsData("sprt_emp_ageHier")

MapHierarchiesToData(z, list(age = ageHier, geo = geoDimList, year = yearFormula))

MapHierarchiesToData(data.frame(f = c("A", "B", "C", "D", "E", "A")), list(f = 
       c("AB = A + B", "AC = A + C", "CD = C + D", "ABCD = AB + CD")))
       
}
