% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlabColon.R
\name{matlabColon}
\alias{matlabColon}
\title{Simulate Matlab's ':'}
\usage{
matlabColon(from, to)
}
\arguments{
\item{from}{numeric. The start value}

\item{to}{numeric. The end value.}
}
\value{
A numeric vector, possibly empty.
}
\description{
A function to simulate Matlab's ':' operator
}
\details{
matlabColon(a,b) returns a:b (R's version) unless a > b, in which case it returns numeric(0).
}
\note{
The function is stolen from own package ffmanova.
}
\examples{
identical(3:5, matlabColon(3, 5)) ## => TRUE
3:1 ## => 3 2 1
matlabColon(3, 1) ## => numeric(0)
}
\seealso{
\code{\link{seq}}
}
\author{
Bjørn-Helge Mevik
}
