% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix2list.R
\name{Matrix2list}
\alias{Matrix2list}
\alias{Matrix2listInt}
\title{Convert matrix to sparse list}
\usage{
Matrix2list(x)

Matrix2listInt(x)
}
\arguments{
\item{x}{Input matrix}
}
\value{
A two-element list: List of row numbers (r) and a list of numeric or integer values (x)
}
\description{
Convert matrix to sparse list
}
\details{
Within the function, the input matrix is first converted to a dgTMatrix matrix (Matrix package).
}
\note{
\code{Matrix2listInt} convers the values to integers by \code{as.integer}
and no checking is performed. Thus, zeros are possible.
}
\examples{
m = matrix(c(0.5, 1.1, 3.14, 0, 0, 0, 0, 4, 5), 3, 3)
Matrix2list(m)
Matrix2listInt(m)
}
\author{
Øyvind Langsrud
}
