% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixPaste.R
\name{MatrixPaste}
\alias{MatrixPaste}
\alias{MatrixPaste1}
\title{Combining columns of a matrix}
\usage{
MatrixPaste(x, sep = "_", forceCharacter = FALSE, stringEmpty = " ")

MatrixPaste1(x, stringEmpty = "1")
}
\arguments{
\item{x}{Matrix or vector}

\item{sep}{String used to combine columns}

\item{forceCharacter}{When FALSE single column input will keep to original class in output.}

\item{stringEmpty}{String used when input is empty (can be set to NULL)}
}
\value{
Character vector or possibly same vector as input
}
\description{
Combining columns of a matrix
}
\details{
Each row in input will be combined to a single string using sep.
}
\examples{
\dontrun{
MatrixPaste(matrix(1:12,3,4))
MatrixPaste(1:5)
MatrixPaste(1:5, forceCharacter=TRUE)
MatrixPaste(matrix(integer(0),3,0))
MatrixPaste(NULL)
}
}
\author{
Øyvind Langsrud
}
\keyword{internal}
