% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrameToMatrix.R
\name{MatrixToDataFrame}
\alias{MatrixToDataFrame}
\title{MatrixToDataFrame}
\usage{
MatrixToDataFrame(x, forceStringsAsFactors = FALSE)
}
\arguments{
\item{x}{matrix}

\item{forceStringsAsFactors}{forceStringsAsFactors}
}
\value{
data frame
}
\description{
MatrixToDataFrame
}
\author{
Øyvind Langsrud
}
\keyword{internal}
