% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_aggregate.R
\name{model_aggregate}
\alias{model_aggregate}
\title{model_aggregate}
\usage{
model_aggregate(
  data,
  sum_vars,
  fun_vars,
  fun = sum,
  hierarchies = NULL,
  formula = NULL,
  dimVar = NULL,
  charVar = NULL,
  pre_aggregate = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{data}

\item{sum_vars}{sum_vars}

\item{fun_vars}{fun_vars}

\item{fun}{fun}

\item{hierarchies}{hierarchies}

\item{formula}{formula}

\item{dimVar}{dimVar}

\item{charVar}{charVar}

\item{pre_aggregate}{pre_aggregate}

\item{verbose}{verbose}

\item{...}{dots}
}
\value{
data frame
}
\description{
model_aggregate
}
\examples{
z <- SSBtoolsData("sprt_emp_withEU")
z$age[z$age == "Y15-29"] <- "young"
z$age[z$age == "Y30-64"] <- "old"
names(z)[names(z) == "ths_per"] <- "ths"
z$y <- 1:18

my_range <- function(x) c(min = min(x), max = max(x))

out <- model_aggregate(z, 
   formula = ~age:year + geo, 
   sum_vars = c("y", "ths"), 
   fun_vars = c(sum = "ths", mean = "y", med = "y", ra = "ths"), 
   fun = c(sum = sum, mean = mean, med = median, ra = my_range))

out
}
