% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMatrix.R
\name{Model_Matrix}
\alias{Model_Matrix}
\title{Overparameterized model matrix}
\usage{
Model_Matrix(
  formula,
  data = NULL,
  mf = model.frame(formula, data = data),
  allFactor = TRUE,
  sparse = FALSE
)
}
\arguments{
\item{formula}{formula}

\item{data}{data frame}

\item{mf}{model frame (alternative input instead of data)}

\item{allFactor}{When TRUE all variables are coerced to factor}

\item{sparse}{When TRUE sparse matrix created by sparse.model.matrix()}
}
\value{
model matrix created via model.matrix() or sparse.model.matrix()
}
\description{
All factor levels included
}
\examples{
  z <- SSBtoolsData("sprt_emp_withEU")
  Model_Matrix(~age*year + geo, z)
}
\keyword{internal}
