% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddLeadingZeros.R
\name{Number}
\alias{Number}
\title{Adding leading zeros}
\usage{
Number(n, width = 3)
}
\arguments{
\item{n}{numeric vector of whole numbers}

\item{width}{width}
}
\value{
Character vector
}
\description{
Adding leading zeros
}
\examples{
Number(1:3)
}
\author{
Øyvind Langsrud
}
