% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumSingleton.R
\name{NumSingleton}
\alias{NumSingleton}
\title{Decoding of \code{singletonMethod}}
\usage{
NumSingleton(singletonMethod)
}
\arguments{
\item{singletonMethod}{String to be decoded}
}
\value{
A list or \code{NULL}
}
\description{
A \code{singletonMethod} starting with \code{"num"} is decoded into internal variables
in \code{\link{GaussSuppression}}
}
\details{
The output variables have the following meaning, when \code{TRUE}:
\itemize{
\item \code{singleton2Primary} (1st character): Published singletons are forced to be primary suppressed.
\item \code{integerUnique} (2nd character): Integer values representing the unique contributors are utilized.
\item \code{sub2Sum} (3rd character, \code{T} or \code{H}): Imaginary primary suppressed cells are made,
which are the sum of some suppressed inner cells and which can be divided into two components.
At least one component is singleton contributor. The other component may be an inner cell.
\item \code{hierarchySearch} (3rd character, \code{H}): As \code{sub2Sum} above.
And in addition, the other component can be any primary suppressed published cell.
}
}
\examples{
unlist(NumSingleton("numTFF"))
unlist(NumSingleton("numFTT"))
unlist(NumSingleton("numTTH"))
}
