% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_term_labels.R
\name{output_term_labels}
\alias{output_term_labels}
\title{Extract vector of term labels from a data.frame}
\usage{
output_term_labels(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
vector of term labels
}
\description{
The data.frame is assumed to be have been constructed with
\code{\link[=ModelMatrix]{ModelMatrix()}} functionality using the \code{formula} parameter.
}
\examples{

out <- model_aggregate(SSBtoolsData("magnitude1"), 
                       formula = ~eu:sector4 + geo * sector2, 
                       sum_vars = "value",
                       avoid_hierarchical = TRUE)
out
term_labels <- output_term_labels(out)
term_labels
cbind(term_labels, out)

}
\seealso{
\code{\link[=formula_term_labels]{formula_term_labels()}}, \code{\link[=tables_by_formulas]{tables_by_formulas()}}
}
