% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rbind0freq.R
\name{Rbind0freq}
\alias{Rbind0freq}
\title{Add zero frequency rows}
\usage{
Rbind0freq(
  data,
  freqName = "freq",
  hierarchical = TRUE,
  varGroups = NULL,
  dimVar = NULL,
  extraVar = TRUE
)
}
\arguments{
\item{data}{data frame}

\item{freqName}{Name of (existing) frequency variable}

\item{hierarchical}{Hierarchical variables treated atomatically when \code{TRUE}}

\item{varGroups}{List of variable groups}

\item{dimVar}{The dimensional variables}

\item{extraVar}{extraVar description}
}
\value{
Extended data frame
}
\description{
Microdata or tabular frequency data is extended
}
\examples{
z <- SSBtoolsData("sprt_emp_withEU")[c(1, 4:6, 8, 11:15), ]
z$age[z$age == "Y15-29"] <- "young"
z$age[z$age == "Y30-64"] <- "old"

Rbind0freq(z[, -4])
Rbind0freq(z, hierarchical = FALSE, dimVar = c("age", "geo", "eu"))
Rbind0freq(z, hierarchical = FALSE, dimVar = c("age", "geo", "eu"), extraVar = "year")
Rbind0freq(z, hierarchical = FALSE, dimVar = c("age", "geo", "eu"), extraVar = FALSE)
Rbind0freq(z, varGroups = list(c("age", "geo", "year"), "eu"))
Rbind0freq(MakeFreq(z[c(1, 1, 1, 2, 2, 3:10), -4]))
Rbind0freq(z, "ths_per")
}
