% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_all_integers.R
\name{table_all_integers}
\alias{table_all_integers}
\title{Table all integers from 1 to n}
\usage{
table_all_integers(x, n)
}
\arguments{
\item{x}{A vector of integers.}

\item{n}{The maximum integer value.}
}
\value{
A 1D array of class \code{"table"} representing the frequency of each integer from 1 to n.
}
\description{
Table all integers from 1 to n
}
\examples{
table_all_integers(c(2, 3, 5, 3, 5, 3), 7)

}
