% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_by_formulas.R
\name{tables_by_formulas}
\alias{tables_by_formulas}
\title{tables_by_formulas}
\usage{
tables_by_formulas(
  data,
  ...,
  table_fun,
  table_formulas,
  substitute_vars = NULL,
  collapse_vars = NULL
)
}
\arguments{
\item{data}{data}

\item{...}{dots}

\item{table_fun}{table_fun}

\item{table_formulas}{table_formulas}

\item{substitute_vars}{substitute_vars}

\item{collapse_vars}{collapse_vars}
}
\value{
A data frame
}
\description{
tables_by_formulas
}
\examples{
tables_by_formulas(SSBtoolsData("magnitude1"),
                   table_fun = model_aggregate, 
                   table_formulas = list(table_1 = ~region * sector2, 
                                         table_2 = ~region1:sector4 - 1, 
                                         table_3 = ~region + sector4 - 1), 
                   substitute_vars = list(region = c("geo", "eu"), region1 = "eu"), 
                   collapse_vars = list(sector = c("sector2", "sector4")), 
                   sum_vars = "value")
}
\seealso{
\code{\link{filter_by_variable}}
}
