% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RowGroups.R
\name{UniqueIntegerSortMatrix}
\alias{UniqueIntegerSortMatrix}
\title{Title UniqueIntegerSortMatrix}
\usage{
UniqueIntegerSortMatrix(x)
}
\arguments{
\item{x}{x}
}
\value{
Matrix
}
\description{
Title UniqueIntegerSortMatrix
}
