% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduce0exact.R
\name{UniqueSeq}
\alias{UniqueSeq}
\title{Sequence within unique values}
\usage{
UniqueSeq(x, sortdata = matrix(1L, length(x), 0))
}
\arguments{
\item{x}{vector}

\item{sortdata}{matrix or vector to determine sequence order}
}
\value{
integer vector
}
\description{
Sequence within unique values
}
\examples{
# 1:4 within A and 1:2 within B
UniqueSeq(c("A", "A", "B", "B", "A", "A"))

# Ordered differently
UniqueSeq(c("A", "A", "B", "B", "A", "A"), c(4, 5, 20, 10, 3, 0))
}
\author{
Øyvind Langsrud
}
