% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_multiple_fun.R
\name{unmatrix}
\alias{unmatrix}
\title{Transform data frame with embedded matrices}
\usage{
unmatrix(data, sep = "_")
}
\arguments{
\item{data}{data frame}

\item{sep}{A character string used when variable names are generated.}
}
\value{
data frame
}
\description{
Transform data frame with embedded matrices
}
\keyword{internal}
