% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hierarchies2Vars.R
\name{Vars2Hierarchies}
\alias{Vars2Hierarchies}
\title{Transform hierarchies coded as Variables to "to-from" format}
\usage{
Vars2Hierarchies(hierarchiesAsVars)
}
\arguments{
\item{hierarchiesAsVars}{As output from \code{\link{Hierarchies2Vars}}}
}
\value{
List of hierarchies
}
\description{
A kind of reverse operation of \code{\link{Hierarchies2Vars}}
}
\examples{

a <- Hierarchies2Vars(list(f1 = 
       c("AB = A + B", "CD = C + D", "AC = A + C", "ABCD = AB + CD")))
a

Vars2Hierarchies(a)

}
