% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchies_as_vars.R
\name{vars_to_hierarchies}
\alias{vars_to_hierarchies}
\title{Transform hierarchies coded as Variables to "to-from" format}
\usage{
vars_to_hierarchies(var_hierarchies)
}
\arguments{
\item{var_hierarchies}{As output from \code{\link{hierarchies_as_vars}}}
}
\value{
List of hierarchies
}
\description{
A kind of reverse operation of \code{\link{hierarchies_as_vars}}
}
\examples{

a <- hierarchies_as_vars(list(f = 
       c("AB = A + B", "CD = C + D", "AC = A + C", "ABCD = AB + CD")))
a

vars_to_hierarchies(a)
}
