% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Div.R
\name{CharacterDataFrame}
\alias{CharacterDataFrame}
\title{CharacterDataFrame}
\usage{
CharacterDataFrame(x)
}
\arguments{
\item{x}{data frame}
}
\value{
data frame
}
\description{
CharacterDataFrame
}
\author{
Øyvind Langsrud
}
\keyword{internal}
