% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{CrossDataDummyHierarchies}
\alias{CrossDataDummyHierarchies}
\title{CrossDataDummyHierarchies}
\usage{
CrossDataDummyHierarchies(dataDummyHierarchies, codeFrames = NULL,
  makeDimnames = FALSE, useMatrixToDataFrame = TRUE,
  removeEmpty = FALSE, verbose = FALSE)
}
\arguments{
\item{dataDummyHierarchies}{dataDummyHierarchies}

\item{codeFrames}{codeFrames}

\item{makeDimnames}{makeDimnames}

\item{removeEmpty}{removeEmpty}

\item{verbose}{Whether to print information during calculations. FALSE is default.}
}
\description{
CrossDataDummyHierarchies
}
\keyword{internal}
