% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{DataDummyHierarchy}
\alias{DataDummyHierarchy}
\title{Create a (signed) dummy matrix for hierarcical mapping of codes in data}
\usage{
DataDummyHierarchy(dataVector, dummyHierarchy)
}
\arguments{
\item{dataVector}{A vector of codes in data}

\item{dummyHierarchy}{Output from DummyHierarchy()}
}
\value{
A sparse matrix.
Column names are taken from dataVector (if non-NULL) and row names are taken from
the row names of dummyHierarchy.
}
\description{
Create a (signed) dummy matrix for hierarcical mapping of codes in data
}
\author{
Øyvind Langsrud
}
\keyword{internal}
