% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimList2Hierarchy.R
\name{DimList2Hierarchy}
\alias{DimList2Hierarchy}
\title{DimList2Hierarchy}
\usage{
DimList2Hierarchy(x)
}
\arguments{
\item{x}{An element of a dimList as in sdcTable}
}
\value{
Data frame with to-from coded hierarchy
}
\description{
From hierarchy/dimList as in sdcTable to to-from coded hierarchy
}
\examples{
# First generate a dimList element 
x <- FindDimLists(SSBtoolsData("sprt_emp_withEU")[, c("geo", "eu")], , total = "Europe")[[1]]

DimList2Hierarchy(x)

}
\author{
Øyvind Langsrud
}
