% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoHierarchies.R
\name{DimList2Hrc}
\alias{DimList2Hrc}
\alias{Hrc2DimList}
\title{DimList2Hrc/Hrc2DimList}
\usage{
DimList2Hrc(dimList)

Hrc2DimList(hrc, total = "Total")
}
\arguments{
\item{dimList}{List of data frames according to the specifications in sdcTable}

\item{hrc}{List of character vectors}

\item{total}{String used to name totals.}
}
\value{
See Arguments
}
\description{
Conversion between hierarchies/dimList as in sdcTable and TauArgus coded hierarchies
}
\examples{
# First generate dimList
dimList <- FindDimLists(SSBtoolsData("sprt_emp_withEU")[, c("geo", "eu", "age")])
hrc <- DimList2Hrc(dimList)
dimList2 <- Hrc2DimList(hrc)
identical(dimList, dimList2)
}
\author{
Øyvind Langsrud
}
