% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DummyApply.R
\name{DummyApply}
\alias{DummyApply}
\title{Apply a function to subsets defined by a dummy matrix}
\usage{
DummyApply(x, y, FUN = sum)
}
\arguments{
\item{x}{A (sparse) dummy matrix}

\item{y}{Vector of input values}

\item{FUN}{A function}
}
\value{
Vector of output values
}
\description{
For each column, \code{i},  of the matrix \code{x} of zeros and ones, the output value is equivalent to \code{FUN(y[x[, i] != 0])}.
}
\details{
With a dummy \code{x} and \code{FUN = sum}, output is equivalent to \code{z = t(x) \%*\% y}.
}
\examples{

z <- SSBtools::SSBtoolsData("sprt_emp_withEU")
z$age[z$age == "Y15-29"] <- "young"
z$age[z$age == "Y30-64"] <- "old"

a <- SSBtools::ModelMatrix(z, formula = ~age + geo, crossTable = TRUE)

cbind(as.data.frame(a$crossTable), 
      sum1 = (t(a$modelMatrix) \%*\% z$ths_per)[,1],
      sum2 = DummyApply(a$modelMatrix, z$ths_per, sum),
       max = DummyApply(a$modelMatrix, z$ths_per, max))

}
