% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forEasySdcTable.R
\name{SortRows}
\alias{SortRows}
\title{Sorting rows}
\usage{
SortRows(m, cols = 1:dim(m)[2], index.return = FALSE)
}
\arguments{
\item{m}{}

\item{cols}{}

\item{index.return}{}
}
\value{
sorted m
}
\description{
Sorting rows
}
\examples{
SortRows(matrix(sample(1:3,15,TRUE),5,3))
}
\author{
Øyvind Langsrud
}
\keyword{internal}
