% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSBtoolsData.R
\name{SSBtoolsData}
\alias{SSBtoolsData}
\title{Function that returns a dataset}
\usage{
SSBtoolsData(dataset)
}
\arguments{
\item{dataset}{Name of data set within the SSBtools package}
}
\value{
data frame
}
\description{
Function that returns a dataset
}
\details{
\strong{FIFA2018ABCD:} A hierarchy table based on
countries within groups A-D in the football championship, 2018 FIFA World Cup.

\strong{sprt_emp:} Employment in sport in thousand persons. Data from Eurostat database.

\strong{sprt_emp_geoHier:}  Country hierarchy for the employment in sport data.

\strong{sprt_emp_ageHier:}  Age hierarchy for the employment in sport data.

\strong{sprt_emp_withEU:} The data set sprt_emp extended with a EU variable.

\strong{my_km2:} Fictitious grid data.

\strong{sosialFiktiv, z1, z1w, z2, z2w, z3, z3w, z3wb:} See \code{\link{sosialFiktiv}}.
}
\examples{
SSBtoolsData("FIFA2018ABCD")
SSBtoolsData("sprt_emp")
SSBtoolsData("sprt_emp_geoHier")
SSBtoolsData("sprt_emp_ageHier")
SSBtoolsData("sprt_emp_withEU")
SSBtoolsData("z1w")
}
\author{
Øyvind Langsrud
}
