% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchyCompute.R
\name{SSBtoolsData}
\alias{SSBtoolsData}
\title{Function that returns a dataset}
\usage{
SSBtoolsData(dataset)
}
\arguments{
\item{dataset}{Name of data set within the SSBtools package}
}
\value{
data frame
}
\description{
Function that returns a dataset
}
\details{
\code{\strong{FIFA2018ABCD:}} A hierarchy table based on
countries within groups A-D in the football championship, 2018 FIFA World Cup.

\code{\strong{sprt_emp:}} Employment in sport in thousand persons. Data from \url{http://ec.europa.eu/eurostat/web/sport/employment-in-sport/data/database}
 
\code{\strong{sprt_emp_geoHier:}}  Country hierarchy for the employment in sport data.

\code{\strong{sprt_emp_ageHier:}}  Age hierarchy for the employment in sport data.
}
\examples{
SSBtoolsData("FIFA2018ABCD")
SSBtoolsData("sprt_emp")
SSBtoolsData("sprt_emp_geoHier")
SSBtoolsData("sprt_emp_ageHier")
}
