% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_multiple_fun.R
\name{unmatrix}
\alias{unmatrix}
\title{Transform data frame with embedded matrices}
\usage{
unmatrix(data, sep = "_")
}
\arguments{
\item{data}{data frame}

\item{sep}{A character string used when variable names are generated.}
}
\value{
data frame
}
\description{
Transform data frame with embedded matrices
}
\examples{
a <- aggregate(1:6, list(rep(1:3, 2)), range)
b <- unmatrix(a)

a
b

dim(a)
dim(b)

names(a)
names(b)

class(a[, 2])
class(b[, 2])
}
\keyword{internal}
